Initial commit. Logger is working with all the main functionality There are tons of optimisation that are required and better test to be written.
At the moment all test require a Elasticsearch instance running on
port 9200. There is a lot of testing required against HTTPS ES instances
and using Kerberos, but this should be a good starting skeleton

From the performance and app consistency point of view same applies,
in some cases logging will lock to commit the buffer changes into
elasticsearch. This should eventually done in a different thread