Should `msg`  also be included in filter fields?
`msg` is not included in logging.Formatter documents, and will be included in the elasticsearch message body, as you're not excluding it here. If you using this handler just for one specific logger statement, it is fine. If you want to apply it to global logger, and there will be various types of `msg` coexisted, then elasticsearch will drop some messages because of that.